+++ rdrf/registry/patients/mixins.py
@@ -1,0 +1,10 @@
+from django.utils.decorators import method_decorator
+from django.contrib.auth.decorators import login_required
+
+
+class LoginRequiredMixin(object):
+
+    @method_decorator(login_required)
+    def dispatch(self, request, *args, **kwargs):
+        return super(LoginRequiredMixin, self).dispatch(
+            request, *args, **kwargs)
