+++ rdrf/registry/patients/patient_view.py
@@ -4,11 +4,12 @@ from django.views.generic.base import View
 
 from rdrf.models.definition.models import Registry
 
+from .mixins import LoginRequiredMixin
 
-class PatientView(View):
+
+class PatientView(View, LoginRequiredMixin):
 
     def get(self, request, registry_code):
-        if request.user.is_authenticated:
-            _ = get_object_or_404(Registry,code=registry_code)
-            patient = request.user.user_object.first()
-            return redirect(reverse("patient_edit", args=[registry_code, patient.id, ]))
+        _ = get_object_or_404(Registry, code=registry_code)
+        patient = request.user.user_object.first()
+        return redirect(reverse("patient_edit", args=[registry_code, patient.id, ]))
